ai-qa-agents/requirements.txt
# AI QA Testing Framework Requirements
# Microsoft Agent Framework (Preview)
agent-framework-azure-ai --pre

# OpenAI Client
openai>=1.40.0

# Azure AI Evaluation SDK
azure-ai-evaluation>=1.0.0
azure-identity>=1.15.0
promptflow>=1.0.0

# Data Processing
pandas>=2.0.0
jsonlines>=4.0.0

# Rich CLI
rich>=13.0.0
typer>=0.9.0
click>=8.1.0

# File Processing
markdown>=3.5.0
pyyaml>=6.0.0
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
python-dotenv>=1.0.0
